Add surface camera platform driver
The Surface Book 1 & 2 use an IPU3-based camera system with OmniVision
camera sensors. The SB2's DSDT is insufficient to allow the IPU3 CIO2
driver to discover the correct camera sensors.

To bridge the gap, hard-code connections between the CIO2 device and the
camera sensors using the software node infrastructure.

The hardcoded values for the software node endpoint are an educated
guess and they currently do not work.

Signed-off-by: Jordan Hand <jorhand@linux.microsoft.com>